requirements.txt
Django
django-grappelli
psycopg2
django-filebrowser
django-extensions
django-debug-toolbar
django-werkzeug
Pillow
coverage
selenium
pytest
python-ldap
